# postman/traffic-controller-API-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/v1/controllers/{controllerId}/commands": {
            "get": {
                "tags": [
                    "Controller Commands"
                ],
                "summary": "Get command history",
                "description": "Returns a full historical list of all commands sent to this controller, sorted by newest first, with pagination",
                "operationId": "getCommandHistory",
                "parameters": [
                    {
                        "name": "controllerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historical audit log retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommandResult"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Controller Commands"
                ],
                "summary": "Send command to device",
                "description": "Sends a specific command to the remote controller and logs the audit trail",
                "operationId": "sendCommand",
                "parameters": [
                    {
                        "name": "controllerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendCommandRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command successfully dispatched and result recorded",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommandResult"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Controller not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommandResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SendCommandRequest": {
                "type": "object",
                "description": "Payload for dispatching a new command to a traffic controller",
                "properties": {
                    "command": {
                        "type": "string",
                        "description": "The command string to be sent to the device"
                    }
                },
                "required": [
                    "command"
                ]
            },
            "CommandResult": {
                "type": "object",
                "description": "Response object representing a recorded command audit trail",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Primary Key of the command",
                        "example": 1234,
                        "readOnly": true
                    },
                    "controllerId": {
                        "type": "string",
                        "description": "Primary Key of the traffic controller",
                        "example": "fd11.z1.downtown.loc",
                        "readOnly": true
                    },
                    "success": {
                        "type": "boolean",
                        "description": "if command wad success or not",
                        "example": "SUCCESS",
                        "readOnly": true
                    },
                    "resultValue": {
                        "type": "string",
                        "description": "The raw return value or acknowledgment from the hardware",
                        "readOnly": true
                    },
                    "deviceTimestamp": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp when the command was processed by the device",
                        "example": "2026-03-07T14:00:00Z",
                        "readOnly": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp when this record created in the DB",
                        "example": "2026-03-07T13:59:58Z",
                        "readOnly": true
                    },
                    "command": {
                        "type": "string",
                        "description": "The specific command instruction dispatched",
                        "enum": [
                            "CHANGE_PROGRAM",
                            "RESET_PROGRAM",
                            "PAUSE_PROGRAM"
                        ],
                        "example": "CHANGE_PROGRAM",
                        "readOnly": true
                    }
                }
            },
...
        }
    }
}